requirements.txt
Appium-Python-Client==5.2.2
attrs==25.3.0
certifi==2025.8.3
exceptiongroup==1.3.0
gherkin-official==29.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parse_type==0.6.6
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
pytest-bdd==8.1.0
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
